+++ lib/factories/index.d.ts
@@ -1,1 +1,2 @@
 export * as react from './react';
+export * as solid from './solid';